+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 4.0.0 (????-??-??)
 
-* DnsPlugins are now just Plugins and can support dns-01 or http-01 based challenges.
+* DnsPlugins are now just Plugins and can support dns-01 or http-01 based challenges. (#124)
   * All existing DNS plugins have been upgraded to the new plugin format. See the README in the plugins folder for details and instructions on how to upgrade your own custom plugins.
 * The `DnsPlugin` parameter is now `Plugin` in `New-PACertificate` and `Submit-ChallengeValidation`.
 * `Publish-DnsChallenge` and `Unpublish-DnsChallenge` are now `Publish-Challenge` and `Unpublish-Challenge`
@@ -11,13 +11,17 @@
   * With a plugin parameter, shows the details for just that plugin
   * With a plugin and `-Help`, shows the plugin's help
   * With a plugin and `-Guide`, opens the default browser to the plugin's online guide
-  * With a plugin and `-Params`, displays the plugin-specific parameter sets
+  * With a plugin and `-Params`, displays the plugin-specific parameter sets (#151)
 * `Get-DnsPluginHelp` has been removed
 * Plugin args are now saved per-order rather than per-account.
   * Unfortunately, this means new orders using the same plugin(s) as a previous order will no longer reuse the previous args.
   * Added `Get-PAPluginArgs` which returns a hashtable with the plugin args associated with the current or specified order. You can use this to retrieve another order's plugin args and use that object with your new order.
   * Pre-4.x plugin args will be automatically migrated to per-order plugin args the first time an account is selected using `Set-PAAccount` or on module load for the last selected account. The old file will be backed up with a ".v3" extension in case you need to revert.
 * Using `Get-PAOrder` with `-Refresh` will no longer throw a terminating error if the ACME server returns an error. It will warn and return the cached copy of the order instead.
+* Certificate orders can now be configured to always generate a new private key and the old parameters for key replacement have been removed. (#181)
+  * `-AlwaysNewKey` has been added to `New-PACertificate`, `New-PAOrder`, and `Set-PAOrder`
+  * `-NewKey` has been removed from `New-PAOrder` and `Submit-Renewal`
+  * `-NewCertKey` has been removed from `New-PACertificate`
 
 ## 3.18.1 (2020-11-12)